--- a/mllib/src/main/scala/org/apache/spark/mllib/regression/IsotonicRegression.scala
+++ b/mllib/src/main/scala/org/apache/spark/mllib/regression/IsotonicRegression.scala
@@ -28,7 +28,7 @@ import org.json4s._
 import org.json4s.JsonDSL._
 import org.json4s.jackson.JsonMethods._
 
-import org.apache.spark.SparkContext
+import org.apache.spark.{SparkContext, SparkException}
 import org.apache.spark.annotation.Since
 import org.apache.spark.api.java.{JavaDoubleRDD, JavaRDD}
 import org.apache.spark.mllib.linalg.{Vector, Vectors}
@@ -345,7 +345,8 @@ class IsotonicRegression private (private var isotonic: Boolean) extends Seriali
     // gives the values for the block. Entries that are not at the start of a block
     // are meaningless.
     val weights: Array[(Double, Double)] = input.map {
-      case (y, _, weight) => (weight, weight * y)
+      case (y, _, weight) => if (weight == 0d) throw new SparkException("Input contains zero weight")
+                             else (weight, weight * y)
     }
 
     // a few convenience functions to make the code more readable