@neggert 2016-10-04
Could someone please take a look at this? @mengxr maybe? Looks like you merged the initial implementation.