@jkbradley 2016-12-14
Thanks for pinging about this---it looks important!

One request: Do you know if there is a resource with a clearer statement of this precise algorithm for the L2 norm?  That'd be nice to have.  The best I found so far was "A Fast Scaling Algorithm for Minimizing Separable Convex Functions Subject to Chain Constraints"