@neggert 2016-12-14
Given that the scikit-learn implementation I based this on has changed (https://github.com/scikit-learn/scikit-learn/pull/7444) since I submitted the PR, and @jkbradley has pointed out a reference I hadn't seen, I'd like to take a little bit of time to see if there's yet a better algorithm.

This one is an improvement over what we had, but as long as we're changing it, we may as well get it right.