@srowen 2016-12-15
If it's correct and faster, I tend to favor merging this. The code looks OK to my understanding and passes existing tests.