@neggert 2016-12-15
Found another input that triggers non-polynomial time with the code in this PR. I'm again borrowing from scikit-learn. I think this is the case they found that led them to re-write their implementation.

```
    val y = ((0 until length) ++ (-(length - 1) until length) ++ (-(length - 1) to 0)).toArray.map(_.toDouble)
    val x = (1 to y.length).toArray.map(_.toDouble)
```

| Input Length | Time (ns) |
| --: | --: |
| 40 | 2059 |
| 80 | 4604 |
| 160 | 1974269 |
| 320 | 3246603433 |

I'm now working on implementing what's described in the Best papers. This should give O(n), even in the worst case. 

Should I close this and open a new PR with the new algorithm, or just add it here and you can squash when you merge?