@srowen 2016-12-27
@neggert I think this is looking good except for the question about zero weights. If throwing an error is OK to you go for it. The test should be restored, but be updated to expect an exception. Other than that, more tests always help, but there is some OK coverage already