@neggert 2017-01-03
Sorry, I've been on vacation. I'll make the requested changes in the next day or two.