@neggert 2017-01-04
Done. It was surprisingly tricky to test, since the exception is thrown in the executor.