@jkbradley 2017-01-12
@neggert I'll ask @mengxr about the negative weights since he oversaw the original work here.