@jkbradley 2017-01-13
@zapletal-martin  Pinging since you wrote the original PR: There's discussion here about whether IsotonicRegression should support negative weights.  Is there a good reason to?  I haven't seen papers which allow negative weights, though I've only looked at a few.  Thanks!