@neggert 2017-01-13
My thought was that there's nothing in the computation that prevents it (unlike 0 weights, which cause NaNs). Why not err on the side of maximum flexibility to the end user? Just because there's no use case for it now doesn't mean someone won't find one in the future.

Just my 2 cents. I don't feel strongly about it and will defer to commiters.