@neggert 2017-01-16
Looks like the build completed successfully. Somehow Jenkins didn't post the message.