@jkbradley 2017-01-18
Thanks for the updates!  This LGTM, except for deciding about negative weights.

Responding to your comment above, negative weights are just as problematic as 0 weights.  See my comment above "Btw, with negative weights, ..."

I spoke with @mengxr offline, and he does not recall why the original work supported negative weights.  I'll wait a few days before merging this to see if @zapletal-martin can check out the issue.