@srowen 2017-01-23
@jkbradley I'll merge if there are no other thoughts here. Seems OK to me too