@jkbradley 2017-01-23
Sounds good.  I'll run fresh tests before merging to be safe though.