@jkbradley 2017-01-23
Merging with master
Thanks!